requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
PyJWT
Werkzeug
scikit-learn
